fix: namespace changed from Sinch.Numbers.EventDestination to Sinch.Numbers.EventDestinations (plural), folder renamed to match, migration guide updated
