+++ src/Sinch/Numbers/Numbers.cs
@@ -13,7 +13,7 @@ using Sinch.Numbers.Available;
 using Sinch.Numbers.Available.List;
 using Sinch.Numbers.Available.Rent;
 using Sinch.Numbers.Available.RentAny;
-using Sinch.Numbers.EventDestination;
+using Sinch.Numbers.EventDestinations;
 
 namespace Sinch.Numbers
 {
